[svn] * NumericLiteral.getIntType returns long, not int.

+++ com.yoursway.ide.ui/src/com/yoursway/utils/RubyASTUtils.java
@@ -33,7 +33,7 @@ public class RubyASTUtils {
         if (stringValue != null)
             return stringValue;
         if (value instanceof NumericLiteral)
-            return Integer.valueOf(((NumericLiteral) value).getIntValue());
+            return Long.valueOf(((NumericLiteral) value).getIntValue());
         return null;
     }
